@@ -26,10 +26,9 @@
 
 import Foundation
 
-
 extension Array {
     public func safeObject(at index: Int) -> Element? {
-        if (0..<count).contains(index) {
+        if (0 ..< count).contains(index) {
             return self[index]
         } else {
             return nil
