@@ -28,15 +28,11 @@ import Foundation
 
 extension DispatchQueue: TiercelCompatible {}
 extension TiercelWrapper where Base: DispatchQueue {
-
-    public static func executeOnMain(_ block: @escaping ()->()) {
+    public static func executeOnMain(_ block: @escaping () -> Void) {
         if Thread.isMainThread {
             block()
         } else {
-            DispatchQueue.main.async {
-                block()
-            }
+            DispatchQueue.main.async(execute: block)
         }
     }
-    
 }
