@@ -26,17 +26,20 @@
 
 import Foundation
 
-
 extension Double: TiercelCompatible {}
+
+fileprivate let globalFormatter = {
+    let formatter = DateFormatter()
+    formatter.dateFormat = "yyyy-MM-dd HH:mm:ss"
+    return formatter
+}()
+
 extension TiercelWrapper where Base == Double {
     /// 返回 yyyy-MM-dd HH:mm:ss格式的字符串
     ///
     /// - Returns:
     public func convertTimeToDateString() -> String {
         let date = Date(timeIntervalSince1970: base)
-        let formatter = DateFormatter()
-        formatter.dateFormat = "yyyy-MM-dd HH:mm:ss"
-        return formatter.string(from: date)
+        return globalFormatter.string(from: date)
     }
-    
 }
