@@ -27,15 +27,13 @@
 import Foundation
 
 extension URLSession {
-    
     /// 把有bug的resumeData修复，然后创建task
     ///
     /// - Parameter resumeData:
     /// - Returns:
     internal func correctedDownloadTask(withResumeData resumeData: Data) -> URLSessionDownloadTask {
-        
         let task = downloadTask(withResumeData: resumeData)
-        
+
         if let resumeDictionary = ResumeDataHelper.getResumeDictionary(resumeData) {
             if task.originalRequest == nil, let originalReqData = resumeDictionary[ResumeDataHelper.originalRequestKey] as? Data, let originalRequest = NSKeyedUnarchiver.unarchiveObject(with: originalReqData) as? NSURLRequest {
                 task.setValue(originalRequest, forKey: "originalRequest")
@@ -44,7 +42,7 @@ extension URLSession {
                 task.setValue(currentRequest, forKey: "currentRequest")
             }
         }
-        
+
         return task
     }
 }
